pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.3.2",
    "fastapi>=0.115.12",
    "uvicorn>=0.34.3",
]
